Add possibility to set more highlighting options
The plugin sets only `fg` values for cterm and gui highlight config. This makes it impossible to easily set highlighting of those parenthesis without manually adding configuration for each `rainbow_pXXX` level.

The way it works is the same as for the fgs. I.e. if one wants to add `bold` to *all* parens-levels it is sufficient to add an array containing only `'bold'`.